Abstracted out calulate total function to it's own utils file for cleaner code. Added caluclate all totals function for line graph data population.
